requirements.txt
SQLAlchemy==2.0.32
tabulate==0.9.0
PyYAML==6.0.2
PyMySQL==1.1.1
python-dotenv==1.0.1